requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
future==0.18.2
geographiclib==1.50
geopy==1.21.0
idna==2.8
multidict==4.7.4
pycodestyle==2.5.0
pycparser==2.19
python-dotenv==0.11.0
python-telegram-bot==12.4.2
requests==2.22.0
six==1.14.0
telepot==12.7
tornado==6.0.3
urllib3==1.25.8
yarl==1.4.2
